refactor(ptt): extract overlay router module
Issue: #77

Why:
Move Client Overlay routing policy out of app.rs so Session routing, stale sequence filtering, mismatched Session drops, Overlay warnings, output hints, and Injection completion routing have a named Module with better Locality and a focused test surface.

What changed:
- Added parakeet-ptt/src/overlay_router.rs for OverlayEvent, OverlaySink, runtime sink construction, OverlayRouter, and router-local tests.
- Updated the Client runtime loop to use the new router Module and keep parent-focus capture independent of router internals.
- Updated main.rs to build the runtime Overlay sink through the new Module.

Compatibility:
No Daemon protocol, Helper, Profile, Stream path, Seal path, Injection timeout, or runtime behavior changes are intended.

Verification (#77):
- targeted: cargo test overlay_router --manifest-path Cargo.toml (from parakeet-ptt/) -> PASSED
- targeted: cargo test overlay --manifest-path Cargo.toml (from parakeet-ptt/) -> PASSED
- targeted: cargo test app_injection_report_error_routes_failure_classification --manifest-path Cargo.toml (from parakeet-ptt/) -> PASSED
- targeted: cargo test mixed_stream_enqueues_exactly_one_final_result --manifest-path Cargo.toml (from parakeet-ptt/) -> PASSED
- regression: N/A, pure refactor with existing behavior coverage and new router-local stale/mismatch tests
- full suite: cargo test --manifest-path parakeet-ptt/Cargo.toml -> PASSED
- lint: cargo clippy --manifest-path parakeet-ptt/Cargo.toml --all-targets --all-features -- -D warnings -> PASSED
- types: cargo check --manifest-path parakeet-ptt/Cargo.toml --all-targets --all-features -> PASSED
- dead-code: N/A, no repo dead-code gate discovered for Rust
- build: cargo build --manifest-path parakeet-ptt/Cargo.toml --bins -> PASSED
- pre-commit: prek run --files parakeet-ptt/src/app.rs parakeet-ptt/src/main.rs parakeet-ptt/src/overlay_router.rs -> PASSED; prek run --all-files -> PASSED
- static/security: N/A, no changed security-sensitive surface and no separate Rust security gate discovered
- migrations: N/A
- CLI/script smoke: N/A, no CLI/script behavior changed
- UI render: N/A, no Overlay renderer UI drawing behavior changed
- destructive/negative: N/A
- review: coderabbit_review_watch.py local --base main -> clean, findings=0, summary=.git/coderabbit-review/20260519T113852Z/status.json (codex-fallback: no)
- tests added/expanded: parakeet-ptt/src/overlay_router.rs
- preexisting failures left: none

Independent review:
CodeRabbit local watcher completed clean with findings=0. PR gate will run after the PR exists.

Deferred follow-ups:
None.
